# huff0: Speed up compression of short blocks