@klauspost 2023-01-20
Wonderful! Never took the time to optimize the table building. Great stuff!